Implement notifications + "mute/unmute chat" menu
    - Added icons: Pinned and muted chats have icons. It's from Twemoji font, I've added
    credits in the README.md and source code of ChatlistItem.qml. I prefer to rely on .PNG for this
    one because icon might not be installed or exist on the system, and we avoid
    the black box problem.
    - Modified CMakeLists: added KNotifications as a required library
    - Modified build.sh: I cannot comprehend how the install(FILES)
    function from cmake works, so I used good old install from GNU coreutils.
    It copies usr/ structure and needed files (.desktop and .notifyrc) for
    KNotifications to work.
    - Aesthetic changes: context menu is shorter now due to the text
    being state-aware. You can now also mute a chat so you won't be notified
    on incoming messages.

+++ context.cpp
@@ -282,7 +282,7 @@ Context::getDraft(uint32_t chatId)
 }
 
 bool
-Context::setChatMuteDuration(uint32_t chatId, int64_t duration)
+Context::setChatMuteDuration(uint32_t chatId, int duration)
 {
     return dc_set_chat_mute_duration(m_context, chatId, duration);
 }
